refactor: 예외 클래스를 exception 패키지로 분리
