+++ src/main/java/nextstep/qna/exception/ForbiddenException.java
@@ -1,6 +1,6 @@
-package nextstep.qna;
+package nextstep.qna.exception;
 
-public class ForbiddenException extends RuntimeException{
+public class ForbiddenException extends RuntimeException {
     public ForbiddenException() {
     }
