+++ src/main/java/nextstep/qna/exception/UnAuthenticationException.java
@@ -1,4 +1,4 @@
-package nextstep.qna;
+package nextstep.qna.exception;
 
 public class UnAuthenticationException extends Exception {
     private static final long serialVersionUID = 1L;
