+++ src/main/java/nextstep/users/domain/NsUser.java
@@ -1,6 +1,6 @@
 package nextstep.users.domain;
 
-import nextstep.qna.UnAuthorizedException;
+import nextstep.qna.exception.UnAuthorizedException;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
